purge only jobs that had ever some status - prevents to purge grey jobs without status (i.e. with UserTag only) - such greyjob will dacay in LB forever?

+++ org.glite.lb.server/src/srv_purge.c
@@ -325,7 +325,7 @@ int edg_wll_PurgeServer(edg_wll_Context ctx,const edg_wll_PurgeRequest *request,
 					goto abort; 
 				}
 
-				if (now-stat.lastUpdateTime.tv_sec > timeout[stat.state] && !check_strict_jobid(ctx,job))
+				if (stat.lastUpdateTime.tv_sec && now-stat.lastUpdateTime.tv_sec > timeout[stat.state] && !check_strict_jobid(ctx,job))
 				{
 					if (purge_one(ctx,job,dumpfile,request->flags&EDG_WLL_PURGE_REALLY_PURGE,ctx->isProxy)) {
 						edg_wll_FreeStatus(&stat);
